requirements.txt
pillow==10.3.0
requests==2.32.3
urllib3==2.2.2
